requirements.txt
Flask==2.0.2
psycopg2-binary==2.9.3
pgvector==0.1.6
numpy==1.21.4
plotly==5.5.0
networkx==2.6.3
python-dotenv==0.19.2
scipy==1.7.3
scikit-learn==1.0.2
